+++ frontend/src/metabase/querying/parameters/components/DateQuarterYearWidget/DateQuarterYearWidget.tsx
@@ -3,8 +3,10 @@ import { match } from "ts-pattern";
 
 import { QuarterYearPicker } from "metabase/querying/filters/components/QuarterYearPicker";
 import type { QuarterYearPickerValue } from "metabase/querying/filters/types";
-import { serializeDateParameterValue } from "metabase/querying/parameters/utils/dates";
-import { normalizeDateParameterValue } from "metabase/querying/parameters/utils/normalize";
+import {
+  deserializeDateParameterValue,
+  serializeDateParameterValue,
+} from "metabase/querying/parameters/utils/parsing";
 import type { ParameterValueOrArray } from "metabase-types/api";
 
 type DateQuarterYearPickerProps = {
@@ -28,7 +30,7 @@ export function DateQuarterYearWidget({
 function getPickerValue(
   value: ParameterValueOrArray | null | undefined,
 ): QuarterYearPickerValue | undefined {
-  return match(normalizeDateParameterValue(value))
+  return match(deserializeDateParameterValue(value))
     .returnType<QuarterYearPickerValue | undefined>()
     .with({ type: "quarter" }, value => value)
     .otherwise(() => undefined);
