+++ frontend/src/metabase/querying/parameters/components/DateSingleWidget/DateSingleWidget.tsx
@@ -7,8 +7,10 @@ import {
   SingleDatePicker,
   type SingleDatePickerValue,
 } from "metabase/querying/filters/components/DatePicker/SpecificDatePicker/SingleDatePicker";
-import { serializeDateParameterValue } from "metabase/querying/parameters/utils/dates";
-import { normalizeDateParameterValue } from "metabase/querying/parameters/utils/normalize";
+import {
+  deserializeDateParameterValue,
+  serializeDateParameterValue,
+} from "metabase/querying/parameters/utils/parsing";
 import type { ParameterValueOrArray } from "metabase-types/api";
 
 type DateSingleWidgetProps = {
@@ -44,7 +46,7 @@ export function DateSingleWidget({
 function getPickerValue(
   value: ParameterValueOrArray | null | undefined,
 ): SingleDatePickerValue | undefined {
-  return match(normalizeDateParameterValue(value))
+  return match(deserializeDateParameterValue(value))
     .returnType<SingleDatePickerValue | undefined>()
     .with({ type: "specific", operator: "=" }, ({ values, hasTime }) => ({
       date: values[0],
